requirements.txt
Flask==3.0.0
google-generativeai==0.8.3
Pillow==10.4.0
pdf2image==1.17.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
Werkzeug==3.0.1
